Add a new class Utility and test case S13_TC06
